Address PR review feedback for #15
Must Fix:
- Add EditError handling in main() to gracefully handle edit conflicts
- Add missing newline at end of requirements.txt

Should Fix:
- Reorganize imports in bot.py (separate standard/third-party/local)
- Add integration tests for edit conflict handling with baserevid

Nitpicks:
- Improve comment clarity in filters.py (remove informal "bodge")

Questions Addressed:
- Remove redundant dryrun_page computation from config.py
- Split TopicDetector tests into dedicated test_topic.py file
- Update type hint for mediawiki_param_nowiki to accept Optional[str]
- Make python-dotenv optional (wrap in try-except for production)

All 125 tests passing.
